+++ InternshipProject.DAL/Repositories/HallRepository.cs
@@ -14,7 +14,7 @@ public class HallRepository : IRepository<Hall>
         _db = db;
     }
 
-    public IEnumerable<Hall>? GetAll()
+    public IEnumerable<Hall> GetAll()
     {
         return _db.Halls;
     }
@@ -42,10 +42,6 @@ public class HallRepository : IRepository<Hall>
     public void Delete(int id)
     {
         var hall = _db.Halls.Find(id);
-        if (hall != null)
-        {
-            _db.Halls.Remove(hall);
-            
-        }
+        if (hall != null) _db.Halls.Remove(hall);
     }
 }
