+++ InternshipProject.DAL/Repositories/SectionRepository.cs
@@ -2,14 +2,14 @@
 using InternshipProject.DAL.Entities;
 using InternshipProject.DAL.Interfaces;
 using Microsoft.EntityFrameworkCore;
-using InvalidOperationException = System.InvalidOperationException;
 
 namespace InternshipProject.DAL.Repositories;
 
 public class SectionRepository : IRepository<Section>
 {
     private readonly CinemaContext _db;
-    public IEnumerable<Section>? GetAll()
+
+    public IEnumerable<Section> GetAll()
     {
         return _db.Sections;
     }
@@ -37,9 +37,6 @@ public class SectionRepository : IRepository<Section>
     public void Delete(int id)
     {
         var section = _db.Sections.Find(id);
-        if (section != null)
-        {
-            _db.Sections.Remove(section);
-        }
+        if (section != null) _db.Sections.Remove(section);
     }
 }
